refactor: simplify web app setup
