+++ vite.config.ts
@@ -1,5 +1,5 @@
 import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react'
+import react from '@vitejs/plugin-react-swc'
 import path from 'path'
 
 // https://vitejs.dev/config/
@@ -7,11 +7,14 @@ export default defineConfig({
   plugins: [react()],
   resolve: {
     alias: {
-      '@': path.resolve(__dirname, './src')
+      '@': path.resolve(__dirname, './src'),
+      'react-native': 'react-native-web',
+      'expo-constants': 'expo-constants/build/index.web.js'
     }
   },
   define: {
-    global: {}
+    global: {},
+    __DEV__: process.env.NODE_ENV !== 'production'
   },
   server: {
     port: 3000
